src/requirements.txt
gym
gymnasium
torch
numpy
plotly
dash
dash-core-components
dash-html-components
matplotlib
pyyaml
absl-py
tensorboard
tensorflow
tensorflow-estimator
Werkzeug
zipp
python-dotenv